+++ src/Movim/Controller/Front.php
@@ -74,11 +74,6 @@ class Front extends Base
         $c = $this->loadController($request);
 
         if (is_callable([$c, 'load'])) {
-            // Useful for the daemon
-            if (php_sapi_name() != 'cli' && $request == 'login') {
-                file_put_contents(CACHE_PATH . 'baseuri', BASE_URI);
-            }
-
             $c->name = $request;
             $c->load();