+++ CHANGELOG.md
@@ -30,6 +30,7 @@ v0.32 (master)
 * Fix #1473 Fix and refactor the Pubsub services discovery form
 * Add a templater worker and start to offload some widget calls to it
 * Refactor and simplify the internal events system, enforce Packet everywhere
+* Remove the necessity to launch the Login page to cache the BASE_URI for the daemon launch
 
 v0.31
 ---------------------------